config/conexion.py
import streamlit as st
import mysql.connector
from datetime import datetime

def conectar_bd():
    try:
        conn = mysql.connector.connect(
            host="tu_host",
            user="tu_usuario", 
            password="tu_contraseña",
            database="tu_base_datos"
        )
        return conn
    except mysql.connector.Error as e:
        st.error(f"Error de conexión: {e}")
        return None

def main():
    st.set_page_config(
        page_title="Sistema de Grupos de Ahorro y Crédito",
        page_icon="💰",
        layout="wide"
    )
    
    st.title("🏦 Sistema de Gestión de Grupos de Ahorro y Crédito")
    
    # Menú principal basado en los módulos de tu BD
    menu = st.sidebar.selectbox(
        "Menú Principal",
        ["Inicio", "Gestión de Grupos", "Socios", "Sesiones", 
         "Ahorros", "Préstamos", "Caja", "Reportes", "Cierres de Ciclo"]
    )
    
    if menu == "Inicio":
        mostrar_inicio()
    elif menu == "Gestión de Grupos":
        gestion_grupos()
    elif menu == "Socios":
        gestion_socios()
    # ... y así con los demás módulos

def mostrar_inicio():
    st.header("Dashboard Principal")
    # Aquí irán métricas y resumen general

def gestion_grupos():
    st.header("👥 Gestión de Grupos")
    # CRUD de grupos, directiva, reglas

def gestion_socios():
    st.header("👤 Gestión de Socios")
    # CRUD de socios, asignación a grupos

# ... más funciones para cada módulo

if __name__ == "__main__":
    main()